+++ OlimpiaBusiness/Interfaces/Exercise1/IDigit.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace OlimpiaBusiness.Implementation.Exercise1
 {
@@ -13,6 +14,6 @@ namespace OlimpiaBusiness.Implementation.Exercise1
         /// </summary>
         /// <param name="number">Numero entero el cual va a ser validado</param>
         /// <returns>Retorna un bool indicando si es multiplo o no.</returns>
-        bool IsMultipleOfThree(int number);
+        bool IsMultipleOfThree(BigInteger number);
     }
 }
